Fix importpath in Go linker -X ldflag
This commit fixes the importpath to fully identify the `commands`
package so that the GitCommit var can be populated at build time.

This will fix binaries build by the `build_redist.sh` script, with the
Brew built release requiring the same update in the homebrew-core
formula (ie bumping the version on the command line isn't sufficient
this time).

Signed-off-by: John McCabe <john@johnmccabe.net>