String trailing spaces/newlines from url in url-ping
Prior to this commit attempting to use the `faas-cli invoke` command to
trigger the `url-ping` function would fail due to the presence of a
trailing newline when piping to STDIN.

For example, running the following:
```
$ echo "http://yahoo.com" | faas-cli invoke --gateway
http://127.0.0.1:8080 --name url-ping
Reading from STDIN - hit (Control + D) to stop.
Handle this -> http://yahoo.com

Timed out trying to reach URL.
```
Note the extra line after the `Handle this` output.

This commit uses `rstrip` to trim the url before using it, and results
in the url passed to requests being valid.
```
$ echo "http://yahoo.com" | faas-cli invoke --gateway
http://127.0.0.1:8080 --name url-ping
Reading from STDIN - hit (Control + D) to stop.
Handle this -> http://yahoo.com
http://yahoo.com => 200
```

Signed-off-by: John McCabe <john@johnmccabe.net>