Suppress STDIN reading prompt when STDIN is piped
Prior to this commit the `Reading from STDIN..` message would be printed
to stderr regardless of whether the `invoke` had input piped to stdin,
or it was being read from the terminal.

This commit checks whether the input to stdin is from the terminal or a
pipe and suppresses the message for the pipe case.
```
$ echo "http://yahoo.com" | ./faas-cli invoke --gateway http://127.0.0.1:8080 --name url-ping
Handle this -> http://yahoo.com
http://yahoo.com => 200

$ ./faas-cli invoke --gateway http://127.0.0.1:8080 --name url-ping
Reading from STDIN - hit (Control + D) to stop.
http://google.comHandle this -> http://google.com
http://google.com => 200
```

Signed-off-by: John McCabe <john@johnmccabe.net>