Fix command help texts for new_function
Removes the `http://localhost:8080` from the gateway flag, its not
necessary as the default will be printed in the help output.

Altered the Use: to better represent the flag options, ie 2 choices
between `--list` or with `--name` and `--lang` with an optional
`--gateway`.

Added a new example with the `--gateway` set.

Tweaked the error messages returned when `--name` or `--lang` are
absent.

Signed-off-by: John McCabe <john@johnmccabe.net>