Remove language default and fix command flags
This commit stops using `node` as a default language due to it
conflicting with functions where the `fprocess` is set in the image.

Given the large numbers of languages having a default makes less sense
now.

It also removes some of the copy-paste/redundant/irrelevant flag
definitions that ended up in newly added subcommands.

Signed-off-by: John McCabe <john@johnmccabe.net>