Add Opt-out analytics to CLI
This commit implements #108 adding an opt-out analytics capture using
Google Analytics.

Refer to the analytics.md file for details of what is being captured at
this point and how to opt out from the capture.

*Note* it does not capture function counts etc, pending the merge of
openfaas/faas#315.

@alexellis - you will need to replace the `trackingID` and
`applicationName` in `analytics/google_analytics.go` in order to direct
events to a GA account associated with the OpenFaaS project.

Signed-off-by: John McCabe <john@johnmccabe.net>