Add OPENFAAS_URL to store_deploy
The recently added feature to draw the gateway from an ENV var
to save users having to repeatedly use -g was missed from the
store's deploy feature.  This meant that even if the ENV var
was set the CLI would try to use the default.

This change adds a call to getGatewayURL immediately prior to
heading into deploy image.

Signed-off-by: rgee0 <richard@technologee.co.uk>