Set defaultNetwork to empty value
A recent change to faas-swarm means that if a network value isnt provided
then the provider will try to ascertain the target network by looking for
a label of openfaas=true within the network meta data.

This change will set the CLI defaultNetwork value to blank so that users
who have not used the default network name will still attain desirable
results when using the store & will no longer have to provide a network
over-ride from the CLI.

This is a breaking change and requires 0.7.3 or greater of the gateway.

Signed-off-by: rgee0 <richard@technologee.co.uk>