Add build-arg flag to build command
Enables build-flag to be passed through to the Docker CLI for
use with secrets or build-time configuration.

Fixes #363 and referenced issues.

Tested with plain Dockerfile template and adding an argument in
to be written to a file during a read-stage. That successfully
was captured in a file and read back at run-time showing the
change works. I also added validation for empty k/v or a lack of
separator (=).

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>