Update error message for faas-cli deploy
If the user makes a mistake and issues faas-cli deploy <stack.yml>
instead of faas-cli deploy -f <stack.yml>, the error message  that
is  thrown  can  be  confusing:  "Please provide  a  --image to be
deployed". With this change the error message is more detailed for
the case when the -f is ommited by mistake.

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>