Accomodate spaces in --build-arg value
With the recent addition of --build-arg no provision for the potential for spaces
was made.  This was a problem as the slice that is passed into exec.command() was being
generated via a strings.Split() on space which meant values with spaces were split
across slice elements.

This change alters this behaviour by returning a slice of strings from the buildFlag function
which obviates the need for string.Split further downstream and so removes the potential for
mis-handling of spaces.

Signed-off-by: rgee0 <richard@technologee.co.uk>