Add SHA checking to get.sh download script
SHA256 checksum files were recently added to the releases page.  This change uses these
in order to check the received binary. A new function has been introduced called checkHash
which checks that shasum is available on the system and if it does goes on to test the
newly received binary against the checksum calculated at build time.  A subshell is invoked
within this function as shasum needs to be called in the same dir aas the binary and this location
can vary depending on the user context.  Using a subshell means the directory switching is transparent
to the user, who remains where they started regardless of the script outcome.

Signed-off-by: Richard Gee <richard@technologee.co.uk>