Replace shorthand status flag with long verion in get.sh
As part of the graceful degradation change use of sha256sum was introduced.
On Ubuntu 16.04 the shorthand flag for status does not exist which results in
a failure when verifying the checksum.  Further, support for -s & the long --status
is inconsistent across distros.  This change replaces the -s option which suppresses
output in favour of redirection to /dev/null.

Signed-off-by: Richard Gee <richard@technologee.co.uk>