Fix registry auth for Gitlab subdir images
This commit fixes the getRegistryAuth function so that it works
correctly with images that exist in subdirs on custom docker registries,
for example Gitlab.

Before this change the following image works as expected:

    registry.gitlab.com/johnmccabe/myfunctions

But a imaged nested under that namespace would fail to find any auth
from the docker cached creds, for example this would fail:

    registry.gitlab.com/johnmccabe/myfunctions/coolfn

This fix address the underlying issue in the registry domain extraction
so that both cases now work, I've tested with private registries on
both Gitlab and Docker Hub successfully.

Have also added a test to catch the nested subdir image case.

Signed-off-by: John McCabe <john@johnmccabe.net>