Push Git SHA tagged image with `--tag` flag
By default `docker push` will push all tagged images with the same
image id so `faas-cli push` will do as well. With this change one
may explicitly point to the image tagged with the latest version.

This follows up the change for `faas-cli build` that generates
image version of type `gittag-gitsha` or `latest-gitsha` if there
is no tag associated with te latest commit.

Note that the version is generated during the execution of
`faas-cli push --tag`. If there are no updates of the Git history
after `faas-cli build --tag` it will generate the same version and
push the proper image. Otherwise it will fail trying to push a not
existing image. It future this should be addressed and resolved by
storing the versions in temporary files

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>