Deploy Git SHA tagged image with `--tag` flag
With  this  change  one may deploy the latest specified version of
the function image, which is of type `gittag-gitsha`  by executing
`faas-cli deploy --tag`.

This  follows up  the change for  `faas-cli build`  that generates
image version of type  `gittag-gitsha` or `latest-gitsha` if there
is no tag associated with te latest commit.

Note  that  the  version  is generated  during  the  execution  of
`faas-cli deploy --tag`.  If  there  are  no  updates  of  the Git
history after `faas-cli build --tag`  it  will  generate  the same
version and deploy the proper image. Otherwise it will fail trying
to push a not existing image.  It  future this should be addressed
and resolved by storing the versions in temporary files

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>