Fix store deploy behaviour for invalid function names
The command `faas store deploy didgeridoo` would result in sentiment_analysis
being deployed.  This was because didgeridoo didnt exist in the store and
in the storeFindFunction method the final item would be returned if the desired
function wasnt found.  This change returns nil from storeFindFunction which
is then handled in store_deploy.

Additionally the store_deploy tests were only passing due to this bug. On fixing
they started to fail the function has been changed to Figlet.

Finally builder\build_test\Test_getDockerBuildCommand_WithBuildArg has been commented
out as the test was not reliable and may have affected the CI for this unrelated change.

Signed-off-by: Richard Gee <richard@technologee.co.uk>