Add HMAC message signatures to invoke command
Change adds the ability to sha1 sign invoke calls with a pre-shared key.
Signing requires  both the header value to store the hash and the key used
to generate the hash so checking for both values is implemented.

Once the hash has been generated and the header value constructed the
header is appended to, and treated the same as, the existing set of headers.

Unit tests added to cover the new methods with the generateSignedHeader tests
both generating and validating the generated values.

Signed-off-by: Richard Gee <richard@technologee.co.uk>