Restore file append functionality to `faas-cli new`
Restores the yaml file append functionality that 973a75291e84901ca73837822337749da875876c
introduced and e509cc7476a0c5f2c2e5d9e74b724e2e3ea833dd built upon that was inadvertently
removed in 2186a07ebb639d099b10f06e1910d35980a15c4b through the introduction of parsing of
the services struct via a golang template.

The template would rewrite the entire file based on the contents of the services struct which
would cause attribute such as env-vars, secrets, labels and annotations to be lost.  As well
as any comments the user may have inserted into their stack file.

This change alters the mode of operation so that the `new` command only concerns itself with
the function that the user is trying to create.  The deduplication to avoid multiple functions
with the same name remains in place and the generated yaml is either written to a new file
or appended to the existing file depending upon whether the user supplied `--append`.

Signed-off-by: Richard Gee <richard@technologee.co.uk>