Add check for file when not in append mode
Reworking of the append function introduced a characteristic whereby
if a handler directory was deleted but the yaml file remained then
the content would be appended to the file even when append was not
set.

This commit adds a check for the file existing if the append mode is
not set and exists with an error, informing the user that the yaml
file already exists.

Signed-off-by: Richard Gee <richard@technologee.co.uk>