Make secrets function specific
Prior to this change secrets would accumulate as each function spec was processed.
This meant that the CLI could potentially hide misconfigurations within the yaml,
which when combined with the non-deterministic map iteration would lead to unpretictable
deployment behaviour.

This change resets the secrets slice on each iteration so that only the secrets associatedwith a function are mounted into that function.

Signed-off-by: Richard Gee <richard@technologee.co.uk>