Add version to stack file
A version is added to the stack YAML files starting at version 1.0. An
additional optional flag is also added.

This was tested by writing unit tests to validate that the faas-cli
adds a proper version to the stack YAML file.

The version flag is tested with unit tests that cover:
- if an empty string is provided in the flag, return error
- if an invalid version is provided in the flag, return error
- if a valid version is provided, success
- if no version is provided, stack file created w/ default version

Functions to parse the stack YAML files were also updated to read
and validate versions. Unit tests were written for this that cover:
- if no version provided, assume default version and no error
- if an invalid version is provided, return error
- if a valid version is provided, no error

Signed-off-by: Kevin Turcios <kevin_turcios@outlook.com>