Generate CRD from store images
This feature adds the support for deploying to the operator using
a CRD from the Function Store. A new flag --from-store can be
passed that allows for matching upon the name. It means we can now
easily deploy figlet and similar functions to the operator through
CRDs.

Tested with new unit tests and manual testing to generate a CRD
positive / negative testing using the strings figlet / filgets.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>