Initial auth command
Opens browser for user to authenticate using the implicit grant
flow. https://oauth.net/2/grant-types/implicit/

A temporary HTTP server is started to receive the token, which
is only available via JavaScript in a URL fragment. The embedded
JavaScript is used to send the token to the server to be
returned to the user.

The patch includes a temporary update to the list command so
that a --token argument can be passed in instead of basic-auth.

Tested e2e with my oidc-plugin and Auth0.

Special thanks to @LucasRoesler for his input.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>