Add client_credentials to auth command
This patch adds the client_credentials flow so that it can be
used in CI/CD processes for deploying to the OpenFaaS gateway
without using an interactive flow.

Tested e2e with Auth0 and a Machine2Machine App type.
Tested that there is no regression in UI / gateway code flow, or
the implicit flow.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>