Fix `template store list` command
An issue existed whereby arm64 templates would not display in the `template list` command.

The cause was located within the `formatTemplatesOutput()` function which was seemingly doing some pre-filtering to ascertain whether the output should return empty if no templates matching the supplied pattern were found.  Within this function a comparison was being performed of the template.Platform string length and the range index of the current slice.  This meant that as arm64 didnt occur within the first 6/7 elements the function would return and empty string.

Fix:

Move the call to filterTemplate() into the formatTemplatesOutput() function and then test for a non-zero slice length before proceeding.  As the templates slice is filtered higher up this removes the need to call in to filterTemplate() within the two output functions, which consequently also removes the need to pass platform into them.

Signed-off-by: Richard Gee <richard@technologee.co.uk>