Refactor exec and buildformat to improve type safety and clearer imports
**What**
- Create and move exec commands to a new exec package, this will make it
easier to reuse in the future "install" command
- Move git related commands to the versioncontrol package to keep the
git related commands together and easy to find.
- Implement flags interfaces for the BuildFormat so that the flags
package can set and validate this value directly and then we can
directly pass this value around to improve type safety in the throughout
- Renames the `tag` variable to `tagFormat` to improve readability, the
value is not directly the tag value and this could be confusing before.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>