Restore getTemplateURL() in template pull
Rework in #668 removed the call to getTemplateURL() which set `repository` to a value based on a precedence order.  This meant that an argument needed to be passed to `faas template pull` in order for it to complete without error, which had implications for pulling of the standard templates using the bare command.

This commit re-introduces the line which determines what value should be used for the repository before the code heads off into pullTemplate()

Signed-off-by: Richard Gee <richard@technologee.co.uk>