Add tests for npm install script
This change relates only to the npm install script
It rearranges the script in order to make the individual
functions testable without affecting the actual installation.
Added unit tests for each of the cases in the getSuffix function
to ensure that the results will remain as expected.
No testing framework dependency is used in an effort to reduce
the size of the installation as the faas-cli uses the npm
installation process and we don't want to pull more dependencies
than absolutely necessary

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>