Download kubeseal via faas-cli cloud seal command
Closes: #722 and closes: #724 by @utsavanand2

This is the combination of the approach taken by @utsavanand2 and
the guidance offered by @rgee0 and @LucasRoesler combined into
a single PR. It also uses the GitHub HTTP website instead of the
GitHub API - the GitHub API has a rate-limit of 50 requests per
hour and is likely to get exhausted quickly, so this approach
removes the issue in #724.

Given that there is only one executable needed at this stage, the
download command idea was removed in favour of a simple flag. The
flag will be updated in the README.md and documentation after
the merge and release.

Tested with --download and --download-version with an override
on MacOS.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>