Depreciate generic "faas" provider for "openfaas"
The openfaas provider was introduced as a default over 8 months
ago, so anyone still using the generic and legacy term "faas"
should update their stack.yml file. Simply replacing "faas" with
"openfaas" is sufficient, all files created over the past
8 months+ should already have this text.

This is the final step of the depreciation of the legacy
provider "faas" and has no bearing on functionality.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>