Return an error if kubeseal release is not found
Prviously the kubeseal download code would just download any
web page without checking if we got a OK result code back, this
lead to a file that was a htlm 404 page being downloaded. Now
we check for a 200 OK, and if we don't get that we return an
error.

This has been tested by building for ARM, where the kubeseal
binary is not available, and running on a Raspberry Pi.
We got the new error on Rpi, and the binary downloaded
fine on amd64

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>