Update new command for new handler path
The new command needs to decide whether to find its templated
handler code in the function folder or a custom folder.

Tested with the old cli and a template modified to use "src".
The result was an empty directory, I then updated faas-cli with
this patch and got the files populated as expected: i.e.
handler.js and package.json

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>