Update struct for OpenFaaS Function store
Looks like the store yaml is different to the expected structure in the
cli for generating the yaml spec from the function store definitions.
There was a missing wrapper sruct that took the version and the list of
functions.

This has been tested with go build && ./faas-cli generate --from-store
nodeinfo

This generated yaml when there was an error before
Signed-off-by: Alistair Hey <alistair@heyal.co.uk>