Don't warn about insecure connections to localhost
This message isn't strictly required or beneficial when
connecting to 127.0.0.1 or localhost, where certificates from
Let's Encrypt are not applicable. If using Kubernetes or SSH
port forwarding, then the connection is encrypted already.

Tested by creating an SSH tunnel to a faasd instance, and
seeing no warning, then connecting to the same faasd instance
with its IP addresses, and continuing to see the error.

The warning message has been updated since SSL is not
applicable to HTTPS currently.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>