Add sort order to generate command
The output from faas-cli generate is not deterministic.  The function
detail is stored in a map which is by design non-deterministic.
This behaviour is undesirable as generating from the same stack will
result in a different file, hence makes tracking diffs, particularly
in git, quite a challenge.

This change seeks to introduce a pre-processing stage where a slice
of function names is created and a rudimentary sort applied. This slice
is then used to iterate through the map using the function name.

Signed-off-by: Richard Gee <richard@technologee.co.uk>