Add function config to ignore paths during shrinkwrap
This change handles an `ignore_paths` configuration array that prevents
specified folders and files from being copied into the handler build
folder.

This is useful for node applications where the process of copying a
handler's node_modules folder can take a long time. In such case, the
node_modules folder is not needed as it is ignored during the Docker
build process anyway.

This change also applies to file paths that are copied from the
configurations.copy array.

Signed-off-by: Tate Barber <tate.barber@gmail.com>