Add multiple env vars to describe test table
As env vars are stored in a `map[string]string` the print output of this
through describe is non-deterministic.  This change introduces a test case
that will cause `TestDescribeOuput()` to become flakey as the four env vars
could be emitted in a different order each time the test is run.

This will be addressed in a future commit which will make env var ordering
consistent and repeatable.

Signed-off-by: Richard Gee <richard@technologee.co.uk>