Add a soft failure if version lookup fails
Currently if an error is returned from the Version lookup against GitHub
the CLI will exit with an error.  This is unnecessary as the version lookup
is an informational item.

This change removes the error in favour of a message to the terminal to
inform the user that a version lookup wasnt possible.

Signed-off-by: Richard Gee <richard@technologee.co.uk>