feat: add run command for local testing
Improve the DX by simplifying the process to run a functional locally.
This makes it easier to quickly build and test a function.

For simplicity, this only supports running functions defined in stack
files.

AE: Convert to local-run, improve error when no arguments are
given. Print port to clarify to user where the server is listening.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>
Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>