Add fprocess to generated CRDs and fix indentation
Indentation changed to four spaces when upgrading to the v3
YAML paser, this reverts to 2 spaces.

The fprocess variable in the store was missing from generated
function YAML files with faas-cli generate --from-store - this
has now been fixed and an env will be emitted when there is
no existing conflicting definition in the store's manifest.

Tested before/after - env now can be deployed, previously it
couldn't be.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>