Support --pull for build/publish commands
The --pull flag is supported by Docker and forces a pull
attempt to be made for base images for the case that there
is a newer version available.

The main use-case for this is when teams are publishing
production or internal images directly from their own machines
which accumulate older images.

This is not needed in CI pipelines which start with an empty
library, so always pull the base images fresh from source.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>