Pull templates from store when not present
Implements a user-experience changes:

faas-cli new - when the templates folder is not present the
language will be looked up in the template store using the
default URL or an overriden one.

faas-cli publish - when no template folder is available then
a pull will be carried out if a configuration section is
present with a template given within in

Filtering - whilst there is no way to avoid pulling all the
templates wihtin a repository, only the requested/needed ones
are expanded by commands like faas-cli new.

Tested mainly with existing-unit tests, faas-cli new / publish
with and without a templates folder.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>