Update faas-cli new to use stack.yaml
Updates faas-cli new to create a stack.yaml file instead of
a YAML file named after the function. This is to help users
discover that stack.yaml does not require a --yaml or -f
flag which can become cumbersome.

When a stack.yaml or the file specified via -f already exists
the new command will now prompt users to use --append

Convert to SHA256 signatures for the invoke command

This is a convenience option for invoking functions along with
a signature, if a function accepts one. They should all have been
moved to sha256 now if they were even using it

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>